refactor(meta): simplify iceberg v3 job control
