--- src/meta/src/barrier/checkpoint/independent_job/iceberg_v3_job/barrier_control.rs
@@ -1,61 +1,0 @@
-// Copyright 2026 RisingWave Labs
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use risingwave_common::id::JobId;
-use risingwave_common::metrics::{LabelGuardedHistogram, LabelGuardedIntGauge};
-use risingwave_common::util::epoch::EpochPair;
-
-use crate::barrier::partial_graph::PartialGraphStat;
-use crate::rpc::metrics::GLOBAL_META_METRICS;
-
-pub(super) struct IcebergV3BarrierStats {
-    consuming_snapshot_barrier_latency: LabelGuardedHistogram,
-    consuming_log_store_barrier_latency: LabelGuardedHistogram,
-    inflight_barrier_num: LabelGuardedIntGauge,
-
-    snapshot_epoch: u64,
-}
-
-impl IcebergV3BarrierStats {
-    pub(super) fn new(job_id: JobId, snapshot_epoch: u64) -> Self {
-        let table_id_str = format!("{}", job_id);
-        Self {
-            snapshot_epoch,
-            consuming_snapshot_barrier_latency: GLOBAL_META_METRICS
-                .snapshot_backfill_barrier_latency
-                .with_guarded_label_values(&[table_id_str.as_str(), "consuming_snapshot"]),
-            consuming_log_store_barrier_latency: GLOBAL_META_METRICS
-                .snapshot_backfill_barrier_latency
-                .with_guarded_label_values(&[table_id_str.as_str(), "consuming_log_store"]),
-            inflight_barrier_num: GLOBAL_META_METRICS
-                .snapshot_backfill_inflight_barrier_num
-                .with_guarded_label_values(&[&table_id_str]),
-        }
-    }
-}
-
-impl PartialGraphStat for IcebergV3BarrierStats {
-    fn observe_barrier_latency(&self, epoch: EpochPair, barrier_latency_secs: f64) {
-        let barrier_latency_metrics = if epoch.prev < self.snapshot_epoch {
-            &self.consuming_snapshot_barrier_latency
-        } else {
-            &self.consuming_log_store_barrier_latency
-        };
-        barrier_latency_metrics.observe(barrier_latency_secs);
-    }
-
-    fn observe_barrier_num(&self, inflight_barrier_num: usize, _collected_barrier_num: usize) {
-        self.inflight_barrier_num.set(inflight_barrier_num as _);
-    }
-}
